removed try catch from room instrumentation, as twp makes it useless added new Room classes (generated through ksp) to VisitorTest
