--- examples/android-instrumentation-sample/new room test/FavoritesDao/FavoritesDao_Impl.java
@@ -1,227 +1,0 @@
-package io.sentry.android.instrumentation.lib.data;
-
-import android.database.Cursor;
-import androidx.annotation.NonNull;
-import androidx.room.CoroutinesRoom;
-import androidx.room.EntityDeletionOrUpdateAdapter;
-import androidx.room.EntityInsertionAdapter;
-import androidx.room.RoomDatabase;
-import androidx.room.RoomSQLiteQuery;
-import androidx.room.SharedSQLiteStatement;
-import androidx.room.util.CursorUtil;
-import androidx.room.util.DBUtil;
-import androidx.sqlite.db.SupportSQLiteStatement;
-import java.lang.Class;
-import java.lang.Exception;
-import java.lang.Integer;
-import java.lang.Object;
-import java.lang.Override;
-import java.lang.String;
-import java.lang.SuppressWarnings;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.concurrent.Callable;
-import javax.annotation.processing.Generated;
-import kotlin.Unit;
-import kotlin.coroutines.Continuation;
-import kotlinx.coroutines.flow.Flow;
-
-@Generated("androidx.room.RoomProcessor")
-@SuppressWarnings({"unchecked", "deprecation"})
-public final class FavoritesDao_Impl extends FavoritesDao {
-  private final RoomDatabase __db;
-
-  private final EntityInsertionAdapter<Favorite> __insertionAdapterOfFavorite;
-
-  private final EntityDeletionOrUpdateAdapter<Favorite> __deletionAdapterOfFavorite;
-
-  private final EntityDeletionOrUpdateAdapter<Favorite> __updateAdapterOfFavorite;
-
-  private final SharedSQLiteStatement __preparedStmtOfCount;
-
-  public FavoritesDao_Impl(@NonNull final RoomDatabase __db) {
-    this.__db = __db;
-    this.__insertionAdapterOfFavorite = new EntityInsertionAdapter<Favorite>(__db) {
-      @Override
-      @NonNull
-      protected String createQuery() {
-        return "INSERT OR REPLACE INTO `Favorite` (`FavoriteId`) VALUES (?)";
-      }
-
-      @Override
-      protected void bind(@NonNull final SupportSQLiteStatement statement,
-          @NonNull final Favorite entity) {
-        statement.bindLong(1, entity.getId());
-      }
-    };
-    this.__deletionAdapterOfFavorite = new EntityDeletionOrUpdateAdapter<Favorite>(__db) {
-      @Override
-      @NonNull
-      protected String createQuery() {
-        return "DELETE FROM `Favorite` WHERE `FavoriteId` = ?";
-      }
-
-      @Override
-      protected void bind(@NonNull final SupportSQLiteStatement statement,
-          @NonNull final Favorite entity) {
-        statement.bindLong(1, entity.getId());
-      }
-    };
-    this.__updateAdapterOfFavorite = new EntityDeletionOrUpdateAdapter<Favorite>(__db) {
-      @Override
-      @NonNull
-      protected String createQuery() {
-        return "UPDATE OR REPLACE `Favorite` SET `FavoriteId` = ? WHERE `FavoriteId` = ?";
-      }
-
-      @Override
-      protected void bind(@NonNull final SupportSQLiteStatement statement,
-          @NonNull final Favorite entity) {
-        statement.bindLong(1, entity.getId());
-        statement.bindLong(2, entity.getId());
-      }
-    };
-    this.__preparedStmtOfCount = new SharedSQLiteStatement(__db) {
-      @Override
-      @NonNull
-      public String createQuery() {
-        final String _query = "DELETE FROM Favorite";
-        return _query;
-      }
-    };
-  }
-
-  @Override
-  public Object insert(final Favorite favorite, final Continuation<? super Unit> $completion) {
-    return CoroutinesRoom.execute(__db, true, new Callable<Unit>() {
-      @Override
-      @NonNull
-      public Unit call() throws Exception {
-        __db.beginTransaction();
-        try {
-          __insertionAdapterOfFavorite.insert(favorite);
-          __db.setTransactionSuccessful();
-          return Unit.INSTANCE;
-        } finally {
-          __db.endTransaction();
-        }
-      }
-    }, $completion);
-  }
-
-  @Override
-  public Object insertAll(final Favorite[] favorites,
-      final Continuation<? super Unit> $completion) {
-    return CoroutinesRoom.execute(__db, true, new Callable<Unit>() {
-      @Override
-      @NonNull
-      public Unit call() throws Exception {
-        __db.beginTransaction();
-        try {
-          __insertionAdapterOfFavorite.insert(favorites);
-          __db.setTransactionSuccessful();
-          return Unit.INSTANCE;
-        } finally {
-          __db.endTransaction();
-        }
-      }
-    }, $completion);
-  }
-
-  @Override
-  public Object delete(final Favorite favorite, final Continuation<? super Unit> $completion) {
-    return CoroutinesRoom.execute(__db, true, new Callable<Unit>() {
-      @Override
-      @NonNull
-      public Unit call() throws Exception {
-        __db.beginTransaction();
-        try {
-          __deletionAdapterOfFavorite.handle(favorite);
-          __db.setTransactionSuccessful();
-          return Unit.INSTANCE;
-        } finally {
-          __db.endTransaction();
-        }
-      }
-    }, $completion);
-  }
-
-  @Override
-  public Object update(final Favorite favorite, final Continuation<? super Unit> $completion) {
-    return CoroutinesRoom.execute(__db, true, new Callable<Unit>() {
-      @Override
-      @NonNull
-      public Unit call() throws Exception {
-        __db.beginTransaction();
-        try {
-          __updateAdapterOfFavorite.handle(favorite);
-          __db.setTransactionSuccessful();
-          return Unit.INSTANCE;
-        } finally {
-          __db.endTransaction();
-        }
-      }
-    }, $completion);
-  }
-
-  @Override
-  public Object count(final Continuation<? super Integer> $completion) {
-    return CoroutinesRoom.execute(__db, true, new Callable<Integer>() {
-      @Override
-      @NonNull
-      public Integer call() throws Exception {
-        final SupportSQLiteStatement _stmt = __preparedStmtOfCount.acquire();
-        try {
-          __db.beginTransaction();
-          try {
-            final Integer _result = _stmt.executeUpdateDelete();
-            __db.setTransactionSuccessful();
-            return _result;
-          } finally {
-            __db.endTransaction();
-          }
-        } finally {
-          __preparedStmtOfCount.release(_stmt);
-        }
-      }
-    }, $completion);
-  }
-
-  @Override
-  public Flow<List<Favorite>> all() {
-    final String _sql = "SELECT * FROM Favorite ORDER BY FavoriteId DESC";
-    final RoomSQLiteQuery _statement = RoomSQLiteQuery.acquire(_sql, 0);
-    return CoroutinesRoom.createFlow(__db, false, new String[] {"Favorite"}, new Callable<List<Favorite>>() {
-      @Override
-      @NonNull
-      public List<Favorite> call() throws Exception {
-        final Cursor _cursor = DBUtil.query(__db, _statement, false, null);
-        try {
-          final int _cursorIndexOfId = CursorUtil.getColumnIndexOrThrow(_cursor, "FavoriteId");
-          final List<Favorite> _result = new ArrayList<Favorite>(_cursor.getCount());
-          while (_cursor.moveToNext()) {
-            final Favorite _item;
-            final long _tmpId;
-            _tmpId = _cursor.getLong(_cursorIndexOfId);
-            _item = new Favorite(_tmpId);
-            _result.add(_item);
-          }
-          return _result;
-        } finally {
-          _cursor.close();
-        }
-      }
-
-      @Override
-      protected void finalize() {
-        _statement.release();
-      }
-    });
-  }
-
-  @NonNull
-  public static List<Class<?>> getRequiredConverters() {
-    return Collections.emptyList();
-  }
-}
