+++ plugin-build/src/test/resources/testFixtures/instrumentation/androidxRoom/java/delete/DeletionDao_DeleteQuery.class
