+++ plugin-build/src/test/resources/testFixtures/instrumentation/androidxRoom/java/delete/DeletionDao_Impl.class
