+++ plugin-build/src/test/resources/testFixtures/instrumentation/androidxRoom/java/insert/InsertionDao_Impl.class
