+++ plugin-build/src/test/resources/testFixtures/instrumentation/androidxRoom/java/select/SelectDao_FlowSingle.class
