+++ plugin-build/src/test/resources/testFixtures/instrumentation/androidxRoom/java/update/UpdateDao_UpdateQueryAndReturnInteger.class
